+++ src/test/java/io/retel/ariproxy/persistence/plugin/InMemoryPersistenceStore.java
@@ -1,5 +1,6 @@
-package io.retel.ariproxy.boundary.callcontext;
+package io.retel.ariproxy.persistence.plugin;
 
+import io.retel.ariproxy.persistence.PersistenceStore;
 import io.vavr.collection.HashMap;
 import io.vavr.collection.Map;
 import io.vavr.concurrent.Future;
