+++ src/test/resources/application.conf
@@ -36,4 +36,7 @@ service {
     commands-topic = "commandsTopic"
     events-and-responses-topic = "eventsAndResponsesTopic"
   }
+
+  persistence-store = "io.retel.ariproxy.persistence.plugin.InMemoryPersistenceStore"
+
 }
